requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
requests>=2.31.0
beautifulsoup4>=4.12.0
